+++ dis_snek/api/http/http_requests/threads.py
@@ -4,7 +4,7 @@ import discord_typings
 
 from dis_snek.client.const import MISSING, Absent
 from ..route import Route
-from dis_snek.client.utils.converters import timestamp_converter
+from dis_snek.client.utils.attr_converters import timestamp_converter
 
 __all__ = ["ThreadRequests"]
