+++ dis_snek/client/utils/__init__.py
@@ -1,6 +1,6 @@
 from .attr_utils import *
 from .cache import *
-from .converters import *
+from .attr_converters import *
 from .input_utils import *
 from .misc_utils import *
 from .serializer import *
