+++ dis_snek/models/discord/activity.py
@@ -2,7 +2,7 @@ from typing import Optional, List
 
 from dis_snek.client.mixins.serialization import DictSerializationMixin
 from dis_snek.client.utils.attr_utils import define, field
-from dis_snek.client.utils.converters import timestamp_converter, optional
+from dis_snek.client.utils.attr_converters import timestamp_converter, optional
 from dis_snek.client.utils.serializer import dict_filter_none
 from dis_snek.models.discord.emoji import PartialEmoji
 from dis_snek.models.discord.enums import ActivityType, ActivityFlags
